# Optimize `layernom_vectorized_impl` by using adaptive wg selection