--- a/src/ATen/native/xpu/sycl/LayerNormKernels.cpp
+++ b/src/ATen/native/xpu/sycl/LayerNormKernels.cpp
@@ -172,7 +172,6 @@ class LayerNormBackward : public NormBackward<scalar_t, mean_t, weight_t> {
 
 constexpr int vec_size =
     4; // we could make it dependent on dtype, but that would lead to different
-       // results between float and low-p types
 
 // Checks alignment of buffers for using vectorized loads / stores
 template <typename T>
@@ -534,6 +533,13 @@ struct VectorizedLayerNormKernelFunctor
   sycl_local_acc_t<T_ACC> buf_;
 };
 
+int64_t layer_norm_wg_size_select(int64_t max_wg_size, int n) {
+  while (max_wg_size > n && max_wg_size > SIMD) {
+    max_wg_size >>= 1;
+  }
+  return max_wg_size;
+}
+
 template <typename T, typename T_ACC>
 void launch_vectorized_layer_norm_kernel(
     int N,
@@ -546,7 +552,8 @@ void launch_vectorized_layer_norm_kernel(
     T_ACC* mean_data,
     T_ACC* rstd_data) {
   using KernelClass = VectorizedLayerNormKernelFunctor<T, T_ACC>;
-  int64_t wg_size = syclMaxWorkItemsPerSubSlice();
+  auto wg_size =
+      layer_norm_wg_size_select(syclMaxWorkGroupSize<KernelClass>(), N);
   KernelClass kfn(
       N,
       eps,