@EikanWang 2025-05-07
@xytintel , what's the meaning of `(a100-bf16[25088, 16, 24] 1.5ms)`?