@fengyuan14 2025-05-07
Can we change the description? For gut feeling, the optimization reduces more idle computation resources. Occupancy is good? But some of them work, some of them are idle.